requirements.txt
# requirements.txt
python-telegram-bot[ext]==21.0.1
httpx==0.27.0
python-dotenv==1.0.1